Modified roce_config script to do additional configuration associated with UEK5 & two-ports-per-device firmware

+++ roce_config.sh
@@ -8,6 +8,10 @@ NETDEV=""
 W_DCBX=1
 TRUST_MODE=dscp
 PFC_STRING=1,2,3,4,5,6
+CC_FLAG=1
+SET_TOS=0
+MAJOR_VERSION=1
+MINOR_VERSION=1
 
 echo ""
 
@@ -15,7 +19,7 @@ print_usage() {
 #Use this script to configure RoCE on Oracle setups
   echo "Usage:
 	roce_config -i <netdev> [-d <n>] [-t <trust_mode>]
-			[-p <pfc_string>]
+			[-p <pfc_string>] [-q <default_tos>]
 
 Options:
  -i <interface>		enter the interface name(required)
@@ -28,18 +32,62 @@ Options:
  -p <pfc_string>	enter the string of priority lanes to enable pfc for them
 			(default: 1,2,3,4,5,6). This is ignored for dynamic config.
 
+ -q <default_tos>	set the default tos to a value between 0-255. If this option
+			is not used, default tos will remain unchanged.
+
 Example:
 	roce_config -i eth4 -d 0 -t pcp
 "
 }
 
+print_version() {
+	echo "Version: $MAJOR_VERSION.$MINOR_VERSION"
+	echo ""
+}
+
 set_rocev2_default() {
-	cma_roce_mode -d $IBDEV -p $PORT -m 2 > /dev/null
+	echo "RoCE v2" > /sys/kernel/config/rdma_cm/$IBDEV/ports/$PORT/default_roce_mode > /dev/null
 	if [[ $? != 0 ]] ; then
-		>&2 echo " - Setting RoCEv2 as rdma_cm preference failed; Please make sure you installed cma_roce_mode"
+		>&2 echo " - Setting RoCEv2 as rdma_cm preference failed"
 		exit 1
 	else
 		echo " + RoCE v2 is set as default rdma_cm preference"
+	fi
+}
+
+set_tos_mapping() {
+	if [ ! -d "/sys/kernel/config/rdma_cm/$IBDEV/tos_map" ] ; then
+		return
+	fi
+
+	echo 32 > /sys/kernel/config/rdma_cm/$IBDEV/tos_map/tos_map_0
+	if [[ $? != 0 ]] ; then
+		>&2 echo " - Failed to set tos mapping"
+		exit 1
+	fi
+	for i in {1..7}
+	do
+		let "mapping=$i<<5"
+		echo $mapping > /sys/kernel/config/rdma_cm/$IBDEV/tos_map/tos_map_$i
+		if [[ $? != 0 ]] ; then
+			>&2 echo " - Failed to set tos mapping"
+			exit 1
+		fi
+	done
+
+	echo " + Tos mapping is set"
+}
+
+set_deafult_tos() {
+	if [[ $SET_TOS == "0" ]] ; then
+		return
+	fi
+	echo $DEFAULT_TOS > /sys/kernel/config/rdma_cm/$IBDEV/ports/$PORT/default_roce_tos
+	if [[ $? != 0 ]] ; then
+		>&2 echo " - Failed to set default roce tos"
+		exit 1
+	else
+		echo " + Default roce tos is set to $DEFAULT_TOS"
 	fi
 }
 
@@ -54,12 +102,32 @@ config_trust_mode() {
 }
 
 start_lldpad() {
-	service lldpad start > /dev/null
+	if [[ $OS_VERSION == "6" ]] ; then
+		service lldpad start > /dev/null
+	else
+		/bin/systemctl start lldpad.service > /dev/null
+	fi
 	if [[ $? != 0 ]] ; then
 		>&2 echo " - Starting lldpad failed; exiting"
 		exit 1
 	else
 		echo " + Service lldpad is running"
+	fi
+}
+
+#This generic lldpad configuration(not related to RoCE)
+do_lldpad_config() {
+	lldptool set-lldp -i $NETDEV adminStatus=rxtx > /dev/null &&
+	lldptool -T -i $NETDEV -V sysName enableTx=yes > /dev/null &&
+	lldptool -T -i $NETDEV -V portDesc enableTx=yes > /dev/null &&
+	lldptool -T -i $NETDEV -V sysDesc enableTx=yes > /dev/null &&
+	lldptool -T -i $NETDEV -V sysCap enableTx=yes > /dev/null &&
+	lldptool -T -i $NETDEV -V mngAddr enableTx=yes > /dev/null
+	if [[ $? != 0 ]] ; then
+		>&2 echo " - Generic lldpad configuration failed"
+		exit 1
+	else
+		echo " + Finished generic lldpad configuration"
 	fi
 }
 
@@ -89,15 +157,44 @@ enable_pfc_willing() {
 	fi
 }
 
+set_cc_algo_mask() {
+	yes | mstconfig -d 0000:30:00.0 set ROCE_CC_PRIO_MASK_P1=255 ROCE_CC_PRIO_MASK_P2=255 \
+	ROCE_CC_ALGORITHM_P1=ECN ROCE_CC_ALGORITHM_P2=ECN > /dev/null
+	if [[ $? != 0 ]] ; then
+		>&2 echo " - Setting congestion control algo/mask failed"
+		exit 1
+	fi
+}
+
 #This enables congestion control on all priorities, for RP and NP both, 
 #regardless of PFC is enabled on one more priorities.
 enable_congestion_control() {
-	echo 1 > /sys/kernel/debug/mlx5/$PCI_ADDR/cc_params/cc_enable
-	if [[ $? != 0 ]] ; then
-		>&2 echo " - Enabling congestion control failed"
-		exit 1
+	if [ -f "/sys/kernel/debug/mlx5/$PCI_ADDR/cc_params/cc_enable" ] ; then
+		echo 1 > /sys/kernel/debug/mlx5/$PCI_ADDR/cc_params/cc_enable
+		if [[ $? != 0 ]] ; then
+			>&2 echo " - Enabling congestion control failed"
+			exit 1
+		else
+			echo " + Congestion control enabled"
+		fi
 	else
-		echo " + Congestion control enabled"
+		CC_VARS="$(mstconfig -d $PCI_ADDR q | grep ROCE_CC | awk '{print $NF}')"
+		if [[ $? != 0 ]] ; then
+			>&2 echo " - mstconfig query failed"
+			exit 1
+		fi
+		CC_FLAG=1
+		while read -r line; do
+			if [[ $line != "255" && $line != "ECN(0)" ]] ; then
+				CC_FLAG=0
+			fi
+		done <<< "$CC_VARS"
+		if [[ $CC_FLAG == "1" ]] ; then
+			echo " + Congestion control algo/mask are set as expected"
+		else
+			set_cc_algo_mask
+			echo " + Congestion control algo/mask has been changed; Please **REBOOT** to load the new settings"
+		fi
 	fi
 }
 
@@ -133,6 +230,13 @@ case $1 in
 		;;
 	-p )	shift
 		PFC_STRING=$1
+		;;
+	-q )	shift
+		DEFAULT_TOS=$1
+		SET_TOS=1
+		;;
+	-v )	print_version
+		exit
 		;;
 	-h )	print_usage
 		exit
@@ -178,7 +282,48 @@ if [[ $TRUST_MODE != "dscp" && $TRUST_MODE != "pcp" ]] ; then
 	exit 1
 fi
 
+if [[ $SET_TOS == "1" && $DEFAULT_TOS -gt "255" ]] ; then
+	>&2 echo " - Option -q (default tos) can only take values between 0-255"
+	exit 1
+fi
+
+OS_VERSION="$(cat /etc/oracle-release | rev | cut -d" " -f1 | rev | cut -d "." -f 1)"
+if [[ $OS_VERSION != "6" && $OS_VERSION != "7" ]] ; then
+	>&2 echo " - Unexpected OS Version; this script works only for OL6 & OL7"
+	exit 1
+fi
+
+if (! cat /proc/mounts | grep /sys/kernel/config > /dev/null) ; then
+	mount -t configfs none /sys/kernel/config
+	if [[ $? != 0 ]] ; then
+		>&2 echo " - Failed to mount configfs"
+		exit 1
+	fi
+fi
+
+if [ ! -d "/sys/kernel/config/rdma_cm" ] ; then
+	modprobe rdma_cm > /dev/null
+	if [[ $? != 0 ]] ; then
+		>&2 echo " - Failed to load rdma_cm module"
+		exit 1
+	fi
+	if [ ! -d "/sys/kernel/config/rdma_cm" ] ; then
+		>&2 echo " - rdma_cm is missing under /sys/kernel/config"
+		exit 1
+	fi
+fi
+
+if [ ! -d "/sys/kernel/config/rdma_cm/$IBDEV" ] ; then
+	mkdir /sys/kernel/config/rdma_cm/$IBDEV
+	if [[ $? != 0 ]] ; then
+		>&2 echo " - Failed to create /sys/kernel/config/rdma_cm/$IBDEV"
+		exit 1
+	fi
+fi
+
 set_rocev2_default
+set_tos_mapping
+set_deafult_tos
 config_trust_mode
 
 PCI_ADDR="$(ethtool -i $NETDEV | grep "bus-info" | cut -f 2 -d " ")"
@@ -197,6 +342,7 @@ enable_congestion_control
 set_cnp_priority
 
 start_lldpad
+do_lldpad_config
 if [[ $W_DCBX == "0" ]] ; then
 	config_pfc
 else
@@ -204,7 +350,13 @@ else
 fi
 
 echo ""
-echo "Finished configuring \"$NETDEV\" ヽ(•‿•)ノ"
+if [[ $CC_FLAG = "0" ]] ; then
+	>&2 echo "Finished configuring \"$NETDEV\", but needs a *REBOOT*"
+	echo ""
+	exit 1
+else
+	echo "Finished configuring \"$NETDEV\" ヽ(•‿•)ノ"
+fi
 echo ""
 
 ##################################################
